Break out config and fullpath_import.
